requirements.txt
jinja2>=2.10
termcolor>=1.1
pyyaml>=5.1